Refactor the scratchpad handling code to its own file
